requirements.txt
pip==23.2.1
fastapi==0.101.0
orjson==3.9.4
uvicorn==0.23.2
uvloop==0.17.0; sys_platform == "linux" or sys_platform == "darwin"
SQLAlchemy==2.0.20
asyncpg==0.28.0
psycopg2-binary==2.9.7
pydantic-settings ==2.0.2
pydantic ==2.1.1
alembic ==1.11.3
psycopg2==2.9.7
databases[postgresql]==0.6.1


requests == 2.31.0
aiohttp ~= 3.8.5
py7zr == 0.20.6
python-multipart ==0.0.6
fastapi-users ==12.1.2
passlib ==1.7.4
python-jose == 3.3.0


yandex-s3 ==0.1.1
devtools ==0.12.1


pytest==7.4.0
pytest-aiohttp == 1.0.4
pytest-asyncio==0.21.0
httpx==0.24.0
